+++ source/Effect.cpp
@@ -15,7 +15,7 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 
 #include "Effect.h"
 
-#include "Audio.h"
+#include "audio/Audio.h"
 #include "DataNode.h"
 
 using namespace std;
