+++ source/MenuAnimationPanel.cpp
@@ -16,7 +16,7 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 #include "MenuAnimationPanel.h"
 
 #include "Angle.h"
-#include "Audio.h"
+#include "audio/Audio.h"
 #include "Color.h"
 #include "PointerShader.h"
 #include "UI.h"
