+++ source/Visual.cpp
@@ -15,7 +15,7 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 
 #include "Visual.h"
 
-#include "Audio.h"
+#include "audio/Audio.h"
 #include "Effect.h"
 #include "Random.h"
