+++ source/audio/Sound.cpp
@@ -15,7 +15,7 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 
 #include "Sound.h"
 
-#include "File.h"
+#include "../File.h"
 
 #include <AL/al.h>
